removed extra readme.md in ridge_extraction renamed ridge_extraction.py to _extract_ridges.py
