+++ ssqueezepy/__init__.py
@@ -40,7 +40,7 @@ from . import _stft
 from . import _ssq_cwt
 from . import _ssq_stft
 from . import wavelets
-from . import ridge_extraction
+from . import _extract_ridges
 from . import utils
 from . import toolkit
 from . import visuals
@@ -55,7 +55,7 @@ from ._ssq_stft import *
 from ._cwt import *
 from ._stft import *
 from .wavelets import *
-from .ridge_extraction import *
+from ._extract_ridges import *
 from ._test_signals import *
 
 def wavs():
